requirements.txt
huggingface-hub
streamlit
python-dotenv
openai
db-sqlite3
bcrypt
altair